Rename all classes as per interaction pattern.
