+++ test/decryption_test.rb
@@ -1,10 +1,10 @@
 require "./test/test_helper"
-require "./lib/decrypt"
+require "./lib/decryption"
 
-class DecryptTest < Minitest::Test
+class DecryptionTest < Minitest::Test
 
   def setup
-    @decrypt = Decrypt.decode("keder ohulw!", "02715", "040895")
+    @decrypt = Decryption.decode("keder ohulw!", "02715", "040895")
   end
 
   def test_it_accepts_a_date
@@ -20,7 +20,7 @@ class DecryptTest < Minitest::Test
   end
 
   def test_setup_values
-    decrypt_setup = Decrypt.setup("12345", "080196")
+    decrypt_setup = Decryption.setup("12345", "080196")
     assert_equal decrypt_setup[:base], Enigma.generate_characters()
     assert_equal decrypt_setup[:keys], [20, 27, 35, 51]
   end
